+++ picker/App.xaml.cs
@@ -1,0 +1,7 @@
+using System.Windows;
+
+namespace F1SimHubLive.Picker;
+
+public partial class App : Application
+{
+}